chore: move observability setup in a separate module.
